[Feat] move lottie files into a seperated folder.
